增强双系统备份和还原功能  在备份时候： 当系统开启了双系统功能，产生的备份目录 -> 2014-08-10-10-35-31-system0-KOT49H               在没有开启双系统功能时，产生的备份目录 -> 2014-08-10-10-35-31-KOT49H
 在还原时候： 当系统开启了双系统功能，通过函数 -> TWFunc::GetSystemNameFromBackupPath(restore_folder);
              判断恢复目录中是否包含[system0|system1],再删除/data/分区里面相应的目录如 /data/system0

 修改的文件：
            partition.cpp
            twrp-functions.cpp
            twrp-functions.hpp

+++ partition.cpp
@@ -1787,13 +1787,15 @@ bool TWPartition::Restore_Tar(string restore_folder, string Restore_File_System)
 	} else {
 
         if (TDBManager.GetTDBState() && Mount_Point == "/data") {
-            string current_system = TDBManager.GetCurrentSystem();
+            string current_system = TWFunc::GetSystemNameFromBackupPath(restore_folder);
+            if (!current_system.empty()) {
             full_path = Mount_Point + "/" + current_system;
             gui_print("Wiping %s...\n",full_path.c_str());
-            LOGINFO("removedire %s, before restore it",full_path.c_str());
+            LOGINFO("remove dir %s, before restore it",full_path.c_str());
             PartitionManager.Mount_By_Path("/data",true);
             TWFunc::removeDir(full_path,false);
-            mkdir(full_path.c_str(),0755);// reconstruct it
+            }
+
         } else {
 		gui_print("Wiping %s...\n", Display_Name.c_str());
 		if (!Wipe(Restore_File_System))
@@ -1808,11 +1810,14 @@ bool TWPartition::Restore_Tar(string restore_folder, string Restore_File_System)
 
 	Full_FileName = restore_folder + "/" + Backup_FileName;
 	twrpTar tar;
+    /*
     if (full_path != "") {
         tar.setdir(full_path);
     } else {
 	tar.setdir(Backup_Path);
     }
+    */
+    tar.setdir(Backup_Path);
 	tar.setfn(Full_FileName);
 	tar.backup_name = Backup_Name;
 #ifndef TW_EXCLUDE_ENCRYPTED_BACKUPS
